Add clang-tidy checks to ci
Closes #22

+++ src/misc/randn.c
@@ -7,12 +7,12 @@
  * Training: https://swedishembedded.com/training
  */
 
+#include "control/misc.h"
+
+#include <math.h>
 #include <stdbool.h>
 #include <stdlib.h>
 #include <time.h>
-#include <math.h>
-
-#include "control/misc.h"
 
 static float generate_gauss(float mu, float sigma)
 {
@@ -22,23 +22,24 @@ static float generate_gauss(float mu, float sigma)
 
 	if (call == 1) {
 		call = !call;
-		return (mu + sigma * (float)X2);
+		return (mu + sigma * X2);
 	}
 
 	// Compute the uniform norm
 	do {
-		U1 = -1 + ((float)rand() / RAND_MAX) * 2;
-		U2 = -1 + ((float)rand() / RAND_MAX) * 2;
-		W = pow(U1, 2) + pow(U2, 2);
-	} while (W >= 1 || W == 0);
+		U1 = -1.0f + ((float)rand() / (float)RAND_MAX) * 2.0f;
+		U2 = -1.0f + ((float)rand() / (float)RAND_MAX) * 2.0f;
+		W = powf(U1, 2) + powf(U2, 2);
+	} while (W >= 1.0f || W == 0.0f);
 
-	scalar = sqrt((-2 * log(W)) / W);
+	// TODO: potential division by zero
+	scalar = sqrtf((-2.0f * logf(W)) / W);
 	X1 = U1 * scalar;
 	X2 = U2 * scalar;
 
 	call = !call;
 
-	return (mu + sigma * (float)X1);
+	return (mu + sigma * X1);
 }
 
 void randn(float *x, uint16_t length, float mu, float sigma)
